revert: partner applications grid display
