+++ backend/core/package.json
@@ -67,7 +67,7 @@
     "moment": "^2.29.1",
     "nanoid": "^3.1.12",
     "nestjs-throttler-storage-redis": "^0.1.11",
-    "nestjs-typeorm-paginate": "^3.1.3",
+    "nestjs-typeorm-paginate": "^2.2.1",
     "newrelic": "7.5.1",
     "node-polyglot": "^2.4.0",
     "passport": "^0.4.1",
