--- backend/core/src/applications/views/types.ts
@@ -1,10 +1,0 @@
-import { View } from "../../views/base.view"
-
-export enum ApplicationViewEnum {
-  base = "base",
-  partnerList = "partnerList",
-}
-
-export type Views = {
-  [key in ApplicationViewEnum]?: View
-}
