+++ backend/core/src/listings/views/types.ts
@@ -1,10 +1,17 @@
-import { View } from "../../views/base.view"
-
 export enum ListingViewEnum {
   base = "base",
   detail = "detail",
   full = "full",
   partnerList = "partnerList",
+}
+
+export interface View {
+  select?: string[]
+  leftJoins?: {
+    join: string
+    alias: string
+  }[]
+  leftJoinAndSelect?: [string, string][]
 }
 
 export type Views = {
